Backtest integration into Mynt.Core -> Console classes like TableOutputWriter, WriteColored etc. goes to Utility/UtilityConsole
