+++ src/Mynt.Core/Backtester/DatabaseCandleProvider.cs
@@ -8,7 +8,7 @@ using Mynt.Core.Extensions;
 using Mynt.Core.Models;
 using Newtonsoft.Json;
 
-namespace Mynt.Backtester
+namespace Mynt.Core.Backtester
 {
     internal class DatabaseCandleProvider
     {
