+++ src/Mynt.Core/Models/BackTestResult.cs
