+++ src/Mynt.Core/Models/BackTestStrategyResult.cs
