+++ src/Mynt.Core/Models/BackTestTradeResult.cs
